+++ BikeIndex/Model/Authentication/OAuthToken.swift
@@ -70,20 +70,3 @@ extension OAuthToken {
         Date() < expiration
     }
 }
-
-enum Scope: String, CaseIterable, Identifiable {
-    var id: Self { self }
-
-    case readUser = "read_user"
-    case writeUser = "write_user"
-
-    case readBikes = "read_bikes"
-    case writeBikes = "write_bikes"
-}
-
-extension [Scope] {
-    /// Transform this array of `Scope` to be used in an API request for the sign-in page.
-    var queryItem: String {
-        self.map { $0.rawValue }.joined(separator: "+")
-    }
-}
